requirements.txt
Flask==2.3.3
ldap3==2.9.1
python-dotenv==1.0.0
pycryptodome==3.19.0
pymysql==1.1.0
SQLAlchemy==2.0.23
cryptography==41.0.7